fix: rename paths
