+++ src/__tests__/secret-key.test.ts
@@ -1,4 +1,4 @@
-import SEAL from '../throws_wasm_node'
+import SEAL from '../throws_wasm_node_umd'
 import { SEALLibrary } from 'implementation/seal'
 import { EncryptionParameters } from 'implementation/encryption-parameters'
 import { Context } from 'implementation/context'
