+++ src/allows_js_web_es.ts
