+++ src/allows_js_worker_es.ts
