+++ src/allows_wasm_web_es.ts
