+++ src/allows_wasm_worker_es.ts
