+++ src/throws_js_node_es.ts
