+++ src/throws_js_web_es.ts
