+++ src/throws_js_worker_es.ts
