+++ src/throws_wasm_web_es.ts
