+++ src/throws_wasm_worker_es.ts
